# src/swagger.json
{ "openapi": "3.0.0", "info": { "title": "Teste Dvexbit", "description": "API RESTful para gerenciamento de tarefas (to-do list)", "contact": { "email": "lian.mendes26@gmail.com" }, "version": "1.0.0" }, "servers": [ { "url": "http://localhost:3000" } ], "security": [ { "bearerAuth": [] } ], "paths": { "/user": { "post": { "summary": "Cadastro de usuário", "description": "Essa rota é responsável por cadastrar um novo usuário no sistema", "tags": ["User"], "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserInput" }, "examples": { "user": { "value": { "email": "teste@gmail.com", "password": "1234", "firstName": "Teste", "lastName": "da Silva" } } } } } }, "responses": { "201": { "description": "Usuário criado com sucesso", "content": { "application/json": { "schema": { "type": "object", "properties": { "message": { "type": "string" }, "user": { "$ref": "#/components/schemas/User" } } } } } }, "400": { "description": "Verifique se os campos foram preenchidos corretamente" }, "409": { "description": "Já existe um usuário com esse e-mail" } } } }, "/users": { "get": { "summary": "Busca de usuários", "description": "Essa rota é responsável por buscar todos os usuários do sistema", "tags": ["User"], "responses": { "200": { "description": "Usuários encontrados", "content": { "application/json": { "schema": { "type": "object", "properties": { "users": { "type": "array", "items": { "$ref": "#/components/schemas/User" } } } } } } }, "404": { "description": "Nenhum usuário encontrado" } } } }, "/user/{id}": { "get": { "summary": "Busca de usuário por ID", "description": "Essa rota é responsável por buscar um usuário pelo seu ID", "tags": ["User"], "parameters": [ { "in": "path", "name": "id", "schema": { "type": "string" }, "description": "ID do usuário a ser retornado" } ], "responses": { "200": { "description": "Usuário encontrado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } }, "404": { "description": "Usuário não encontrado" } } }, "put": { "summary": "Atualiza usuário", "description": "Essa rota é responsável por atualizar os dados do usuário. Para atualizar a senha, é necessário informar a senha atual.", "tags": ["User"], "parameters": [ { "in": "path", "name": "id", "schema": { "type": "string" }, "description": "ID do usuário a ser atualizado" } ], "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserUpdate" }, "examples": { "user": { "value": { "email": "novoemail@gmail.com", "password": "novaSenha123", "firstName": "NovoNome", "lastName": "NovoSobrenome", "actualPassword": "senhaAtual" } } } } } }, "responses": { "200": { "description": "Usuário atualizado com sucesso", "content": { "application/json": { "schema": { "type": "object", "properties": { "message": { "type": "string" }, "user": { "$ref": "#/components/schemas/UserUpdate" } } } } } }, "400": { "description": "Informe ao menos um valor para atualizar" }, "404": { "description": "Usuário não encontrado" }, "403": { "description": "Senha atual incorreta" }, "409": { "description": "Já existe um usuário com esse e-mail" } } }, "delete": { "summary": "Deleta um usuário", "description": "Essa rota é responsável por deletar um usuário do sistema", "tags": ["User"], "parameters": [ { "in": "path", "name": "id", "schema": { "type": "string" }, "description": "ID do usuário a ser deletado" } ], "responses": { "200": { "description": "Usuário com id {id} deletado com sucesso", "content": { "application/json": { "schema": { "type": "object", "properties": { "message": { "type": "string" } } } } } }, "404": { "description": "Usuário não encontrado" } } } }, "/task": { "post": { "summary": "Criação de tarefa", "description": "Essa rota é responsável por criar uma nova tarefa no sistema", "tags": ["Task"], "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TaskInput" }, "examples": { "task": { "value": { "titulo": "Estudar Python", "descricao": "Estudar operadores", "dataDaAtividade": "2023-10-10", "userId": "123e4567-e89b-12d3-a456-426614174000" } } } } } }, "responses": { "201": { "description": "Tarefa criada com sucesso", "content": { "application/json": { "schema": { "type": "object", "properties": { "message": { "type": "string" }, "task": { "$ref": "#/components/schemas/Task" } } } } } }, "400": { "description": "Erro na criação da tarefa. Verifique se os campos foram preenchidos corretamente ou se a data está no formato YYYY-MM-DD" }, "404": { "description": "Usuário não encontrado" } } } }, "/tasks": { "get": { "summary": "Listagem de tarefas", "description": "Essa rota é responsável por buscar todas as tarefas do sistema", "tags": ["Task"], "responses": { "200": { "description": "Tarefas encontradas", "content": { "application/json": { "schema": { "type": "object", "properties": { "tasks": { "type": "array", "items": { "$ref": "#/components/schemas/Task" } } } } } } }, "404": { "description": "Nenhuma tarefa encontrada" } } } }, "/task/{id}": { "get": { "summary": "Busca de tarefa por ID", "description": "Essa rota é responsável por buscar os detalhes de uma tarefa específica pelo seu ID", "tags": ["Task"], "parameters": [ { "in": "path", "name": "id", "schema": { "type": "string" }, "description": "ID da tarefa a ser retornada" } ], "responses": { "200": { "description": "Tarefa encontrada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Task" } } } }, "404": { "description": "Tarefa não encontrada" } } }, "put": { "summary": "Atualização de tarefa", "description": "Essa rota é responsável por atualizar os dados de uma tarefa existente", "tags": ["Task"], "parameters": [ { "in": "path", "name": "id", "schema": { "type": "string" }, "description": "ID da tarefa a ser atualizada" } ], "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TaskUpdate" }, "examples": { "task": { "value": { "titulo": "Novo título", "descricao": "Nova descrição", "dataDaAtividade": "2023-10-15", "userId": "123e4567-e89b-12d3-a456-426614174000" } } } } } }, "responses": { "200": { "description": "Tarefa atualizada com sucesso", "content": { "application/json": { "schema": { "type": "object", "properties": { "message": { "type": "string" }, "task": { "$ref": "#/components/schemas/Task" } } } } } }, "400": { "description": "Informe ao menos um valor para atualizar ou a data não está no formato YYYY-MM-DD" }, "404": { "description": "Tarefa não encontrada" } } }, "delete": { "summary": "Deleção de tarefa", "description": "Essa rota é responsável por deletar uma tarefa do sistema", "tags": ["Task"], "parameters": [ { "in": "path", "name": "id", "schema": { "type": "string" }, "description": "ID da tarefa a ser deletada" } ], "responses": { "200": { "description": "Tarefa deletada com sucesso", "content": { "application/json": { "schema": { "type": "object", "properties": { "message": { "type": "string" } } } } } }, "404": { "description": "Tarefa não encontrada" } } } }, "/login": { "post": { "summary": "Login de usuário", "description": "Autentica um usuário e retorna um token JWT", "tags": ["Auth"], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginInput" }, "examples": { "login": { "value": { "email": "teste@gmail.com", "password": "1234" } } } } } }, "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Token" } } } }, "400": { "description": "Credênciais inválidas" }, "422": { "description": "Verifique se os campos foram preenchidos" } } } } }, "components": { "schemas": { "User": { "type": "object", "properties": { "id": { "type": "string", "example": "123e4567-e89b-12d3-a456-426614174000" }, "email": { "type": "string", "example": "teste@gmail.com" }, "firstName": { "type": "string", "example": "Teste" }, "lastName": { "type": "string", "example": "da Silva" }, "tasks": { "type":"array", "example": [ { "id": "bc3b624b-7f9d-48a7-8b69-48184c46b637", "titulo": "Estudar programação", "descricao": "Java", "dataDaAtividade": "2025-03-22", "status": "PENDENTE" } ] } } }, "UserInput": { "type": "object", "properties": { "email": { "type": "string", "example": "teste@gmail.com" }, "password": { "type": "string", "example": "1234" }, "firstName": { "type": "string", "example": "Teste" }, "lastName": { "type": "string", "example": "da Silva" } }, "required": ["email", "password", "firstName", "lastName"] }, "UserUpdate": { "type": "object", "properties": { "id": { "type": "string", "example": "123e4567-e89b-12d3-a456-426614174000" }, "email": { "type": "string", "example": "novoemail@gmail.com" }, "password": { "type": "string", "example": "novaSenha123" }, "firstName": { "type": "string", "example": "NovoNome" }, "lastName": { "type": "string", "example": "NovoSobrenome" } } }, "Task": { "type": "object", "properties": { "id": { "type": "string", "example": "123e4567-e89b-12d3-a456-426614174000" }, "titulo": { "type": "string", "example": "Estudar Python" }, "descricao": { "type": "string", "example": "Estudar operadores" }, "dataDaAtividade": { "type": "string", "example": "2023-10-10" }, "status": { "type": "string", "example": "Pendente" }, "userId": { "type": "string", "example": "123e4567-e89b-12d3-a456-426614174000" } } }, "TaskInput": { "type": "object", "properties": { "titulo": { "type": "string", "example": "Estudar Python" }, "descricao": { "type": "string", "example": "Estudar operadores" }, "dataDaAtividade": { "type": "string", "example": "2023-10-10" }, "userId": { "type": "string", "example": "123e4567-e89b-12d3-a456-426614174000" } }, "required": ["titulo", "dataDaAtividade", "userId"] }, "TaskUpdate": { "type": "object", "properties": { "titulo": { "type": "string", "example": "Novo título" }, "descricao": { "type": "string", "example": "Nova descrição" }, "dataDaAtividade": { "type": "string", "example": "2023-10-15" } } }, "LoginInput": { "type": "object", "properties": { "email": { "type": "string", "example": "lian@gmail.com" }, "password": { "type": "string", "example": "1234" } }, "required": ["email", "password"] }, "Token": { "type": "object", "properties": { "token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..." } } } }, "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } } } }